Attempt to roll our own charging but stop before I get carried away
This is actually a real pain in the ass, so we're not going to deal with
it yet. My thoughts are that we should focus on getting something out
the door otherwise I'm going to spend a long time designing a NiMH
charger.

Some problem/open questions in this design:
 - Constant current driver uses a lot of power if we want to support
   fast charging. ~3W of the 6W available.
 - Some kind of switching regulator would be more suited to the task I
   think. BQ24401 seems like a suitable part, relatively cheap.
 - Need an accurate (~1mV according to one source I read) voltage
   sensing circuit if we want to do ∆V/dt termination.

Saving this in this branch to inspire future improvements when we can
prove we need them.